LayoutTests/imported/w3c/web-platform-tests/fetch/sec-metadata/font.tentative.https.sub-expected.txt

1
2
3

FAIL Same-Origin font assert_equals: destination expected (string) "font" but got (undefined) undefined
FAIL Same-Site font assert_equals: destination expected (string) "font" but got (undefined) undefined
FAIL Cross-Site font assert_equals: destination expected (string) "font" but got (undefined) undefined
